- Fixed Language Select styling on Safari - Fixed LoadingSpinner styling on Safari
